change to only checking uint16 and within number range (below 4000)

+++ propertybasedtests/romannumerals.go
@@ -3,7 +3,7 @@ package main
 import "strings"
 
 type RomanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
@@ -23,7 +23,7 @@ var allRomanNumerals = []RomanNumeral{
 	{1, "I"},
 }
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 	// builder is used to efficiently build a string using Write methods. it minimizes memory copying
 	var result strings.Builder
 
@@ -37,8 +37,8 @@ func ConvertToRoman(arabic int) string {
 	return result.String()
 }
 
-func ConvertToArabic(roman string) int {
-	var arabic = 0
+func ConvertToArabic(roman string) uint16 {
+	var arabic = uint16(0)
 	// convenient way to search and trim prefixes
 	for _, numeral := range allRomanNumerals {
 		for strings.HasPrefix(roman, numeral.Symbol) {
